Always ignore tags predating latest in changelog
If there are any missing releases in a changelog, the revision passed
to git log would be wrong, e.g from 0.001 to 0.006, instead of
one release to the next. This commit makes sure any tagged
versions smaller than the the latest changelog entry are skipped.

Also improves handling of tags which don't match version numbers.